# backend/API/openapi.yml
openapi: 3.1.0
info:
  title: Training Management API
  description: API for managing training sessions, students, trainers, schedules, and plans
  version: v1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /sessions:
    get:
      tags:
      - scheduling
      operationId: getSchedule
      parameters:
      - name: startDate
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleResponseDTO'
    post:
      tags:
      - scheduling
      operationId: createOneOffSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneOffSessionCreateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SessionResponseDTO'
...
  /exercises:
    get:
      tags:
      - training
      operationId: findAllExercises
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeInactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageExerciseResponseDTO'
    post:
      tags:
      - training
      operationId: createExercise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExerciseResponseDTO'
  /exercises/{exerciseId}:
    get:
      tags:
      - training
      operationId: findExerciseById
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExerciseResponseDTO'
    put:
      tags:
      - training
      operationId: updateExercise
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExerciseResponseDTO'
    delete:
      tags:
      - training
      operationId: deleteExercise
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
    patch:
      tags:
      - training
      operationId: patchExercise
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchExerciseRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExerciseResponseDTO'
  # ============================================
  # Events Domain
  # ============================================
...
  /sessions/{sessionId}:
    get:
      tags:
      - scheduling
      operationId: getSession
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: trainerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SessionResponseDTO'
    patch:
      tags:
      - scheduling
      operationId: updateSession
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SessionResponseDTO'
  /sessions/{sessionId}/trainer:
    patch:
      tags:
      - scheduling
      operationId: updateSessionTrainer
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionTrainerUpdateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SessionResponseDTO'
  /sessions/{sessionId}/participants:
    post:
      tags:
      - scheduling
      operationId: addSessionParticipant
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionParticipantAddRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponseDTO'
  /sessions/{sessionId}/participants/{studentId}/presence:
    patch:
      tags:
      - scheduling
      operationId: updatePresence
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionParticipantPresenceRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /sessions/{sessionId}/participants/{studentId}/exercises:
    post:
      tags:
      - scheduling
      operationId: addRegisteredParticipantExercise
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantExerciseCreateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: uuid
  /sessions/participants/exercises/{exerciseRecordId}:
    patch:
      tags:
      - scheduling
      operationId: updateRegisteredParticipantExercise
      parameters:
      - name: exerciseRecordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantExerciseUpdateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponseDTO'
    delete:
      tags:
      - scheduling
      operationId: removeRegisteredParticipantExercise
      parameters:
      - name: exerciseRecordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponseDTO'
  /sessions/{sessionId}/cancel:
    patch:
      tags:
      - scheduling
      operationId: cancelOrRestoreSession
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCancelRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SessionResponseDTO'
  # ============================================
  # Plans Domain
  # ============================================
...
  /events/{eventId}/participants/{studentId}:
    post:
      tags:
      - events
      operationId: addParticipant
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
    delete:
      tags:
      - events
      operationId: removeParticipant
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /exercises/{exerciseId}/restore:
    patch:
      tags:
      - training
      operationId: restoreExercise
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExerciseResponseDTO'
...
  /sessions/{sessionId}/participants/{studentId}:
    delete:
      tags:
      - scheduling
      operationId: removeSessionParticipant
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
...
    ExerciseRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 0
        description:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - name
    ExerciseResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
...
    EventParticipantResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatar:
          type: string
        enrolledAt:
          type: string
          format: date-time
        present:
          type: boolean
    EventResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        description:
          type: string
        instructorId:
          type: string
          format: uuid
        instructor:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/EventParticipantResponseDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
...
    Exercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        active:
          type: boolean
        inactive:
          type: boolean
    ParticipantExercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionParticipant:
          $ref: '#/components/schemas/SessionParticipant'
        exerciseId:
          type: string
          format: uuid
        exercise:
          $ref: '#/components/schemas/Exercise'
        setsCompleted:
          type: integer
          format: int32
        repsCompleted:
          type: integer
          format: int32
        weightCompleted:
          type: number
          format: double
        exerciseNotes:
          type: string
        done:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        active:
          type: boolean
    ScheduledSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionSeriesId:
          type: string
          format: uuid
        sessionId:
          type: string
        trainerId:
          type: string
          format: uuid
        trainer:
          $ref: '#/components/schemas/Trainer'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        canceled:
          type: boolean
        seriesName:
          type: string
        notes:
          type: string
        instanceOverride:
          type: boolean
        effectiveFromTimestamp:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        active:
          type: boolean
        participants:
          type: array
          items:
            $ref: '#/components/schemas/SessionParticipant'
    SessionParticipant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        scheduledSession:
          $ref: '#/components/schemas/ScheduledSession'
        studentId:
          type: string
          format: uuid
        student:
          $ref: '#/components/schemas/Student'
        participationType:
          type: string
          enum:
          - INCLUDED
          - EXCLUDED
        attendanceNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        active:
          type: boolean
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantExercise'
        present:
          type: boolean
        included:
          type: boolean
        excluded:
          type: boolean
    SessionUpdateRequestDTO:
      type: object
      properties:
        participants:
          type: array
          items:
            $ref: '#/components/schemas/SessionParticipant'
        notes:
          type: string
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        surname:
          type: string
        gender:
          type: string
          enum:
          - M
          - F
          - O
        birthDate:
          type: string
          format: date
        phone:
          type: string
        profession:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: string
        neighborhood:
          type: string
        cep:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        emergencyContactRelationship:
          type: string
        objectives:
          type: string
        observations:
          type: string
        registrationDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        active:
          type: boolean
        inactive:
          type: boolean
    Trainer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        specialties:
          type: array
          items:
            type: string
        compensationType:
          type: string
          enum:
          - HOURLY
          - MONTHLY
    SessionTrainerUpdateRequestDTO:
      type: object
      properties:
        trainerId:
          type: string
          format: uuid
    SessionParticipantAddRequestDTO:
      type: object
      properties:
        studentId:
          type: string
          format: uuid
    SessionParticipantPresenceRequestDTO:
      type: object
      properties:
        present:
          type: boolean
        notes:
          type: string
    ParticipantExerciseCreateRequestDTO:
      type: object
      properties:
        exerciseId:
          type: string
          format: uuid
        setsCompleted:
          type: integer
          format: int32
        repsCompleted:
          type: integer
          format: int32
        weightCompleted:
          type: number
          format: double
        exerciseNotes:
          type: string
        done:
          type: boolean
    SessionCancelRequestDTO:
      type: object
      properties:
        cancel:
          type: boolean
        reason:
          type: string
    OneOffSessionCreateRequestDTO:
      type: object
      properties:
        seriesName:
          type: string
        trainerId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        notes:
          type: string
    ParticipantExerciseResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        exerciseId:
          type: string
          format: uuid
        exerciseName:
          type: string
        setsCompleted:
          type: integer
          format: int32
        repsCompleted:
          type: integer
          format: int32
        weightCompleted:
          type: number
          format: double
        exerciseNotes:
          type: string
        done:
          type: boolean
    SessionResponseDTO:
      type: object
      properties:
        sessionId:
          type: string
        trainerId:
          type: string
          format: uuid
        trainerName:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        seriesName:
          type: string
        notes:
          type: string
        instanceOverride:
          type: boolean
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentCommitmentResponseDTO'
        canceled:
          type: boolean
        presentCount:
          type: integer
          format: int32
    StudentCommitmentResponseDTO:
      type: object
      properties:
        studentId:
          type: string
          format: uuid
        studentName:
          type: string
        commitmentStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - TENTATIVE
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseResponseDTO'
        participantExercises:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantExerciseResponseDTO'
        present:
          type: boolean
        attendanceNotes:
          type: string
...
    ParticipantExerciseUpdateRequestDTO:
      type: object
      properties:
        setsCompleted:
          type: integer
          format: int32
        repsCompleted:
          type: integer
          format: int32
        weightCompleted:
          type: number
          format: double
        exerciseNotes:
          type: string
        done:
          type: boolean
    PatchExerciseRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 0
        description:
          type: string
          maxLength: 255
          minLength: 0
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    ScheduleResponseDTO:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionResponseDTO'
...
    PageExerciseResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseResponseDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    MessageResponseDTO:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
      - message
      - success
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT